Add flag for using the latest task definition
- This adds the option to use the *latest* task definition that has been created for the family; in
the previous implementation, the only option was to use the currently running task definition as a base.

There are scenarios where a task definition may be created, yet not deployed where this is useful.

This PR will keep the previous implementation as the default so this will not be a breaking change.